move function to shared code
