+++ mgradm/cmd/migrate/shared/shared.go
@@ -5,13 +5,10 @@
 package shared
 
 import (
-	"bytes"
 	"os"
 	"path/filepath"
 
 	"github.com/rs/zerolog/log"
-	"github.com/spf13/viper"
-	"github.com/uyuni-project/uyuni-tools/mgradm/shared/templates"
 	"github.com/uyuni-project/uyuni-tools/shared/utils"
 )
 
@@ -42,62 +39,4 @@ func GetSshPaths() (string, string) {
 	}
 
 	return sshConfigPath, sshKnownhostsPath
-}
-
-func GenerateMigrationScript(sourceFqdn string, kubernetes bool) string {
-	scriptDir, err := os.MkdirTemp("", "mgradm-*")
-	if err != nil {
-		log.Fatal().Err(err).Msgf("Failed to create temporary directory")
-	}
-
-	data := templates.MigrateScriptTemplateData{
-		Volumes:    utils.VOLUMES,
-		SourceFqdn: sourceFqdn,
-		Kubernetes: kubernetes,
-	}
-
-	scriptPath := filepath.Join(scriptDir, "migrate.sh")
-	if err = utils.WriteTemplateToFile(data, scriptPath, 0555, true); err != nil {
-		log.Fatal().Err(err).Msgf("Failed to generate migration script")
-	}
-
-	return scriptDir
-}
-
-func GeneratePgMigrationScript(scriptDir string, oldPgVersion string, newPgVersion string, kubernetes bool) {
-	data := templates.MigratePostgresVersionTemplateData{
-		OldVersion: oldPgVersion,
-		NewVersion: newPgVersion,
-		Kubernetes: kubernetes,
-	}
-
-	scriptPath := filepath.Join(scriptDir, "migrate.sh")
-	if err := utils.WriteTemplateToFile(data, scriptPath, 0555, true); err != nil {
-		log.Fatal().Err(err).Msgf("Failed to generate migration script")
-	}
-}
-
-func GenerateFinalizePostgresMigrationScript(scriptDir string, RunAutotune bool, RunReindex bool, RunSchemaUpdate bool, RunDistroMigration bool, kubernetes bool) {
-	data := templates.FinalizePostgresTemplateData{
-		RunAutotune:        RunAutotune,
-		RunReindex:         RunReindex,
-		RunSchemaUpdate:    RunSchemaUpdate,
-		RunDistroMigration: RunDistroMigration,
-		Kubernetes:         kubernetes,
-	}
-
-	scriptPath := filepath.Join(scriptDir, "migrate.sh")
-	if err := utils.WriteTemplateToFile(data, scriptPath, 0555, true); err != nil {
-		log.Fatal().Err(err).Msgf("Failed to generate migration script")
-	}
-}
-
-func ReadContainerData(scriptDir string) (string, string, string) {
-	data, err := os.ReadFile(filepath.Join(scriptDir, "data"))
-	if err != nil {
-		log.Fatal().Msgf("Failed to read data extracted from source host")
-	}
-	viper.SetConfigType("env")
-	viper.ReadConfig(bytes.NewBuffer(data))
-	return viper.GetString("Timezone"), viper.GetString("old_pg_version"), viper.GetString("new_pg_version")
 }
